fix: 스케쥴러 transactional 해결
